[#220]: Search 페이지 자동완성 - 수정해야 할 것: 새로고침 되는것 어떻게 input 텍스트 유지하는지..
