+++ FrontEnd/rebon-frontend/src/components/Header/SearchModal/index.js
@@ -3,7 +3,8 @@ import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 import Modal, { ModalProvider, BaseModalBackground } from 'styled-react-modal';
 import SearchBar from './SearchBar';
-import axios from 'axios';
+import Tags from '../../../pages/Search/Tags';
+import AutoCompletes from '../../../pages/Search/AutoCompletes';
 import '../../../styles/header-search-modal.css';
 
 export default function SearchModal() {
@@ -28,39 +29,10 @@ export default function SearchModal() {
     });
   }
 
-  const [tag, setTags] = useState([]);
-  useEffect(() => {
-    axios
-      .get('http://3.34.139.61:8080/api/tags')
-      .then((response) => {
-        setTags(response.data);
-        console.log(tag[8]);
-      })
-      .catch((error) => {
-        console.log('error');
-      });
-  }, []);
-
-  // let word = '';
-  // const [word, setWord] = useState('');
-  // const onChangeWord = (e) => {
-  //   // word = e.target.value;
-  //   setWord(e.target.value);
-  //   console.log(word);
-  // };
-
-  const [keyword, setKeyword] = useState([]);
-  useEffect(() => {
-    axios
-      .get(`http://3.34.139.61:8080/api/tags/search?keyword=ㅎㄷㄷ`)
-      .then((response) => {
-        setKeyword(response.data);
-        console.log(keyword);
-      })
-      .catch((error) => {
-        console.log('error');
-      });
-  }, []);
+  const [word, setWord] = useState('');
+  const onChangeWord = (e) => {
+    setWord(e.target.value);
+  };
 
   const FadingBackground = styled(BaseModalBackground)`
     opacity: ${(props) => props.opacity};
@@ -85,24 +57,15 @@ export default function SearchModal() {
           <div className="header-modal-wrapper-search">
             <div className="header-search-bar-wrapper">
               <img src="../../../../image/search-icon.png" alt="header-search-icon" />
-              <input placeholder="가고 싶은 지역을 입력해주세요." />
+              <input placeholder="가고 싶은 지역을 입력해주세요." value={word} onChange={onChangeWord} />
             </div>
           </div>
-
-          <ul className="header-search-list">
-            <li>장성동</li>
-            <li>잔치국수</li>
-            <li>장수동</li>
-          </ul>
-
-          <ul className="header-tag-list">
-            <div>추천 태그로 검색해보세요.</div>
-            {tag.slice(0, 10).map((item) => (
-              <Link to={`/main?name=${item.name}`} state={{ item }}>
-                <li>{item.name}</li>
-              </Link>
-            ))}
-          </ul>
+          <div className="header-autoCom">
+            <AutoCompletes word={word} />
+          </div>
+          <div className="header-tag">
+            <Tags />
+          </div>
         </div>
       </StyledModal>
     </ModalProvider>
@@ -118,4 +81,5 @@ top: -1%;
   border-radius:20px;
   background-color: white;
   opacity: ${(props) => props.opacity};
-  transition : all 0.3s ease-in-out;`;
+  transition : all 0.3s ease-in-out;
+`;
