+++ FrontEnd/rebon-frontend/src/pages/Search/Tags/index.js
@@ -21,7 +21,6 @@ export default function Tags() {
     <ul className="tags-wrapper">
       <div>추천 태그로 검색해보세요.</div>
       {tag.slice(0, 10).map((item) => (
-        //쿼리스트링으로 이렇게 주는게 맞나요오?
         <Link to={`/main?name=${item.name}`} state={{ item }}>
           <li>{item.name}</li>
         </Link>
