+++ FrontEnd/rebon-frontend/src/styles/search-auto-completes.css
@@ -1,0 +1,14 @@
+.auto-list li {
+  width: 95%;
+  margin-left: -7%;
+  padding: 0.4rem 0;
+  border-radius: 10px;
+  color: rgba(0, 0, 0, 0.5);
+  padding-left: 1rem;
+}
+
+.auto-list li:hover {
+  color: #ff6c6c;
+  background-color: #ffeded;
+  border: 0;
+}
