+++ FrontEnd/rebon-frontend/src/styles/search-tags.css
@@ -9,17 +9,18 @@ li {
   font-family: 'Pretendard' !important;
   font-size: 90%;
   padding: 0 2%;
+  margin-top: 1.5rem;
 }
 
 .tags-wrapper div {
-  margin-bottom: 5%;
+  margin: 1rem 0;
   padding-left: 5px;
 }
 
 .tags-wrapper li {
   width: fit-content;
-  padding: 1.5% 3%;
-  margin: 1.5%;
+  padding: 0.4rem 1rem;
+  margin: 0.5rem;
   float: left;
   border: 1px solid #c4c4c4;
   font-weight: bold;
