+++ FrontEnd/rebon-frontend/src/styles/search.css
@@ -18,7 +18,6 @@
 
 .search-content {
   width: 27%;
-  height: 50%;
 }
 
 .search-content .title {
@@ -39,7 +38,7 @@
 .search-content .input-bar {
   width: 95%;
   padding: 1em 0;
-  margin-bottom: 8%;
+  margin-bottom: 0.7rem;
   border-radius: 50px;
   border: 1px solid #c4c4c4;
 }
